try to debug store fault in write syscall

+++ src/utils/console.rs
@@ -1,6 +1,4 @@
-//! SBI console driver, for text output
-
-use crate::sbi::console_putchar;
+use crate::driver::serial::console_putchar;
 use core::fmt::{self, Write};
 
 struct Stdout;
@@ -8,7 +6,7 @@ struct Stdout;
 impl Write for Stdout {
     fn write_str(&mut self, s: &str) -> fmt::Result {
         for c in s.chars() {
-            console_putchar(c as usize);
+            console_putchar(c);
         }
         Ok(())
     }
